처음 체스판의 각 칸을 만드는건 activity에서 수행
